+++ apps/prairielearn/src/lib/assessment.sql
@@ -161,6 +161,46 @@ FROM
 WHERE
   a.id = $assessment_id;
 
+-- BLOCK select_and_lock_assessment_instance_max_points
+SELECT
+  ai.max_points
+FROM
+  assessment_instances AS ai
+WHERE
+  ai.id = $assessment_instance_id
+FOR NO KEY UPDATE OF
+  ai;
+
+-- BLOCK update_assessment_instance_score
+WITH
+  updated_assessment_instances AS (
+    UPDATE assessment_instances AS ai
+    SET
+      points = $points,
+      score_perc = $score_perc,
+      modified_at = now()
+    WHERE
+      ai.id = $assessment_instance_id
+    RETURNING
+      ai.*
+  )
+INSERT INTO
+  assessment_score_logs (
+    assessment_instance_id,
+    auth_user_id,
+    max_points,
+    points,
+    score_perc
+  )
+SELECT
+  ai.id,
+  $authn_user_id,
+  ai.max_points,
+  ai.points,
+  ai.score_perc
+FROM
+  updated_assessment_instances AS ai;
+
 -- BLOCK assessment_instance_log
 WITH
   ai_group_users AS (
@@ -217,6 +257,7 @@ WITH
         NULL::INTEGER AS variant_number,
         NULL::INTEGER AS submission_id,
         NULL::BIGINT AS log_id,
+        NULL::BIGINT AS client_fingerprint_id,
         NULL::JSONB AS data
       FROM
         assessment_instances AS ai
@@ -240,6 +281,7 @@ WITH
         v.number AS variant_number,
         NULL::INTEGER AS submission_id,
         v.id AS log_id,
+        v.client_fingerprint_id AS client_fingerprint_id,
         jsonb_build_object(
           'variant_seed',
           v.variant_seed,
@@ -275,6 +317,7 @@ WITH
         v.number AS variant_number,
         s.id AS submission_id,
         s.id AS log_id,
+        s.client_fingerprint_id AS client_fingerprint_id,
         jsonb_build_object(
           'submitted_answer',
           CASE
@@ -323,6 +366,7 @@ WITH
         v.number AS variant_number,
         gj.id AS submission_id,
         gj.id AS log_id,
+        NULL::BIGINT AS client_fingerprint_id,
         to_jsonb(gj.*) AS data
       FROM
         grading_jobs AS gj
@@ -353,6 +397,7 @@ WITH
         v.number AS variant_number,
         gj.id AS submission_id,
         gj.id AS log_id,
+        NULL::BIGINT AS client_fingerprint_id,
         jsonb_build_object(
           'correct',
           gj.correct,
@@ -423,6 +468,7 @@ WITH
         v.number AS variant_number,
         gj.id AS submission_id,
         gj.id AS log_id,
+        NULL::BIGINT AS client_fingerprint_id,
         jsonb_build_object(
           'correct',
           gj.correct,
@@ -467,6 +513,7 @@ WITH
         v.number AS variant_number,
         NULL::INTEGER AS submission_id,
         qsl.id AS log_id,
+        NULL::BIGINT AS client_fingerprint_id,
         jsonb_build_object(
           'points',
           qsl.points,
@@ -505,6 +552,7 @@ WITH
         NULL::INTEGER AS variant_number,
         NULL::INTEGER AS submission_id,
         asl.id AS log_id,
+        NULL::BIGINT AS client_fingerprint_id,
         jsonb_build_object(
           'points',
           asl.points,
@@ -538,6 +586,7 @@ WITH
         NULL::INTEGER AS variant_number,
         NULL::INTEGER AS submission_id,
         asl.id AS log_id,
+        asl.client_fingerprint_id AS client_fingerprint_id,
         CASE
           WHEN asl.open THEN jsonb_build_object(
             'date_limit',
@@ -583,6 +632,7 @@ WITH
         NULL::INTEGER AS variant_number,
         NULL::INTEGER AS submission_id,
         asl.id AS log_id,
+        NULL::BIGINT AS client_fingerprint_id,
         jsonb_build_object(
           'time_limit',
           format_interval (asl.date_limit - ai.date)
@@ -627,6 +677,7 @@ WITH
         v.number AS variant_number,
         NULL::INTEGER AS submission_id,
         pvl.id AS log_id,
+        pvl.client_fingerprint_id AS client_fingerprint_id,
         NULL::JSONB AS data
       FROM
         user_page_view_logs AS pvl
@@ -654,6 +705,7 @@ WITH
         NULL::INTEGER AS variant_number,
         NULL::INTEGER AS submission_id,
         pvl.id AS log_id,
+        pvl.client_fingerprint_id AS client_fingerprint_id,
         NULL::JSONB AS data
       FROM
         user_page_view_logs AS pvl
@@ -678,6 +730,7 @@ WITH
         NULL::INTEGER AS variant_number,
         NULL::INTEGER AS submission_id,
         gl.id AS log_id,
+        NULL::BIGINT AS client_fingerprint_id,
         jsonb_build_object('user', gu.uid) AS data
       FROM
         assessment_instances AS ai
@@ -687,11 +740,6 @@ WITH
       WHERE
         ai.id = $assessment_instance_id
     )
-    ORDER BY
-      date,
-      event_order,
-      log_id,
-      question_id
   ),
   question_data AS (
     SELECT
@@ -718,15 +766,23 @@ SELECT
   el.variant_number,
   el.submission_id,
   el.data,
+  to_jsonb(cf.*) AS client_fingerprint,
+  NULL AS client_fingerprint_number,
   format_date_full_compact (el.date, ci.display_timezone) AS formatted_date,
   format_date_iso8601 (el.date, ci.display_timezone) AS date_iso8601,
   qd.student_question_number,
   qd.instructor_question_number
 FROM
   event_log AS el
+  LEFT JOIN client_fingerprints AS cf ON (cf.id = el.client_fingerprint_id)
   LEFT JOIN question_data AS qd ON (qd.instance_question_id = el.instance_question_id),
   assessment_instances AS ai
   JOIN assessments AS a ON (a.id = ai.assessment_id)
   JOIN course_instances AS ci ON (ci.id = a.course_instance_id)
 WHERE
-  ai.id = $assessment_instance_id;
+  ai.id = $assessment_instance_id
+ORDER BY
+  el.date,
+  el.event_order,
+  el.log_id,
+  el.question_id;