+++ packages/docker-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @prairielearn/docker-utils
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies [abfd5cc]
+  - @prairielearn/logger@1.0.11
+
+## 2.0.5
+
+### Patch Changes
+
+- Updated dependencies [2da23ab]
+  - @prairielearn/logger@1.0.10
+
 ## 2.0.4
 
 ### Patch Changes